Refactors ScriptQueryBuilder and Parser
Relates to #10217
Closes #12115

This PR is against the query-refactoring branch.
